requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
langchain==0.0.340
langchain-community==0.0.1
langchain-google-genai==1.0.1